+++ app/games/persistence/s3.go
@@ -1,39 +1,29 @@
-package games
+package persistence
 
 import (
 	"errors"
 	"fmt"
-	"time"
 
+	"github.com/cszczepaniak/oh-hell-backend/games"
 	"github.com/cszczepaniak/oh-hell-backend/s3"
 )
-
-type IdGenerator interface {
-	NextId() int64
-}
-
-type TimeStampIdGenerator struct{}
-
-func (ts TimeStampIdGenerator) NextId() int64 {
-	return time.Now().UnixNano()
-}
 
 type S3Persistence struct {
 	KeyFmt      string
 	Client      s3.S3
-	IdGenerator IdGenerator
+	IdGenerator games.IdGenerator
 }
 
 var _ GamePersistence = (*S3Persistence)(nil)
 
-func (sp *S3Persistence) Save(g Game) error {
+func (sp *S3Persistence) Save(g games.Game) error {
 	if g.Id == 0 {
 		return errors.New(`must set ID before saving game`)
 	}
 	return sp.put(g)
 }
 
-func (sp *S3Persistence) Create(g Game) (int64, error) {
+func (sp *S3Persistence) Create(g games.Game) (int64, error) {
 	g.Id = sp.IdGenerator.NextId()
 	err := sp.put(g)
 	if err != nil {
@@ -42,15 +32,15 @@ func (sp *S3Persistence) Create(g Game) (int64, error) {
 	return g.Id, nil
 }
 
-func (sp *S3Persistence) put(g Game) error {
+func (sp *S3Persistence) put(g games.Game) error {
 	return sp.Client.UploadJSON(sp.getGameKey(g.Id), g)
 }
 
-func (sp *S3Persistence) Get(id int64) (Game, error) {
-	var g Game
+func (sp *S3Persistence) Get(id int64) (games.Game, error) {
+	var g games.Game
 	err := sp.Client.DownloadJSON(sp.getGameKey(id), &g)
 	if err != nil {
-		return Game{}, err
+		return games.Game{}, err
 	}
 	return g, nil
 }
